+++ src/runUpdate.ts
@@ -24,6 +24,8 @@ export async function runUpdate(
     showDiff: boolean
 ): Promise<void> {
     const remoteWorkflow = Workflow.fromString(remoteContent);
+    if (!remoteWorkflow.commentLines)
+        remoteWorkflow.commentLines = workflowItem.getCommentLines();
     if (!localContent) {
         console.log(chalk.green("  ✓  created"));
         await workflowItem.setLocal(remoteContent);